chore: improve test case setup
