+++ tests/Unit/BluemPaymentsTest.php
@@ -1,0 +1,18 @@
+<?php
+/*
+ * (c) 2023 - Bluem Plugin Support <pluginsupport@bluem.nl>
+ *
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Bluem\BluemPHP\Tests\Unit;
+
+class BluemPaymentsTest extends BluemTestCase
+{
+
+    public function testCanCreatePaymentRequest(): void
+    {
+        
+    }
+}
